backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python
torch
torchvision
facenet-pytorch
numpy
pandas
Pillow
aiofiles
python-multipart
SQLAlchemy
websockets
